REFACTOR(Domain proj): moved current entities from Data to Domain project
