--- OnlineMinion.Data/Entities/CashAccountSpec.cs
@@ -1,5 +1,0 @@
-using OnlineMinion.Data.Entities.Shared;
-
-namespace OnlineMinion.Data.Entities;
-
-public class CashAccountSpec : BasePaymentSpec { }
