--- OnlineMinion.Data/Entities/TransactionCredit.cs
@@ -1,5 +1,0 @@
-using OnlineMinion.Data.Entities.Shared;
-
-namespace OnlineMinion.Data.Entities;
-
-public class TransactionCredit : BaseTransaction;
