+++ OnlineMinion.Data/OnlineMinion.Data.csproj
@@ -18,7 +18,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\OnlineMinion.Contracts\OnlineMinion.Contracts.csproj"/>
+        <ProjectReference Include="..\OnlineMinion.Domain\OnlineMinion.Domain.csproj"/>
     </ItemGroup>
 
 </Project>
