+++ OnlineMinion.Domain/CashAccountSpec.cs
@@ -1,0 +1,5 @@
+using OnlineMinion.Domain.Shared;
+
+namespace OnlineMinion.Domain;
+
+public class CashAccountSpec : BasePaymentSpec { }
