+++ OnlineMinion.Domain/TransactionCredits/TransactionCredit.cs
@@ -1,0 +1,5 @@
+using OnlineMinion.Domain.Shared;
+
+namespace OnlineMinion.Domain;
+
+public class TransactionCredit : BaseTransaction;
