+++ OnlineMinion.RestApi/AccountSpec/Handlers/CreateAccountSpecReqHlr.cs
@@ -7,8 +7,8 @@ namespace OnlineMinion.RestApi.AccountSpec.Handlers;
 
 [UsedImplicitly]
 internal sealed class CreateAccountSpecReqHlr(OnlineMinionDbContext dbContext)
-    : BaseCreateModelReqHlr<CreateAccountSpecReq, Data.Entities.AccountSpec>(dbContext)
+    : BaseCreateModelReqHlr<CreateAccountSpecReq, Domain.AccountSpec>(dbContext)
 {
-    protected override Data.Entities.AccountSpec ToEntity(CreateAccountSpecReq rq) =>
+    protected override Domain.AccountSpec ToEntity(CreateAccountSpecReq rq) =>
         new(rq.Name, rq.Group, rq.Description);
 }
