+++ OnlineMinion.RestApi/Init/ServicesSetup.cs
@@ -9,8 +9,8 @@ using OnlineMinion.Contracts.AccountSpec.Requests;
 using OnlineMinion.Contracts.PaymentSpec.Requests;
 using OnlineMinion.Contracts.Transactions.Credit.Requests;
 using OnlineMinion.Contracts.Transactions.Debit.Requests;
-using OnlineMinion.Data.Entities;
-using OnlineMinion.Data.Entities.Shared;
+using OnlineMinion.Domain;
+using OnlineMinion.Domain.Shared;
 using OnlineMinion.RestApi.Configuration;
 using OnlineMinion.RestApi.MediatorInfra.Behaviors;
 using OnlineMinion.RestApi.ProblemHandling;
