+++ OnlineMinion.RestApi/PaymentSpec/Handlers/CheckUniqueExistingPaymentSpecReqHlr.cs
@@ -2,7 +2,7 @@ using System.Linq.Expressions;
 using JetBrains.Annotations;
 using OnlineMinion.Contracts.PaymentSpec.Requests;
 using OnlineMinion.Data;
-using OnlineMinion.Data.Entities.Shared;
+using OnlineMinion.Domain.Shared;
 using OnlineMinion.RestApi.Shared.Handlers;
 
 namespace OnlineMinion.RestApi.PaymentSpec.Handlers;
