+++ OnlineMinion.RestApi/PaymentSpec/Handlers/CreatePaymentSpecReqHlr.cs
@@ -1,7 +1,7 @@
 using JetBrains.Annotations;
 using OnlineMinion.Contracts.PaymentSpec.Requests;
 using OnlineMinion.Data;
-using OnlineMinion.Data.Entities.Shared;
+using OnlineMinion.Domain.Shared;
 using OnlineMinion.RestApi.Shared.Handlers;
 
 namespace OnlineMinion.RestApi.PaymentSpec.Handlers;
