+++ OnlineMinion.RestApi/PaymentSpec/Handlers/GetPaymentSpecByIdReqHlr.cs
@@ -2,7 +2,7 @@ using JetBrains.Annotations;
 using OnlineMinion.Contracts.PaymentSpec.Requests;
 using OnlineMinion.Contracts.PaymentSpec.Responses;
 using OnlineMinion.Data;
-using OnlineMinion.Data.Entities.Shared;
+using OnlineMinion.Domain.Shared;
 using OnlineMinion.RestApi.Shared.Handlers;
 
 namespace OnlineMinion.RestApi.PaymentSpec.Handlers;
