+++ OnlineMinion.RestApi/PaymentSpec/Handlers/GetPaymentSpecDescriptorsReqHlr.cs
@@ -2,7 +2,7 @@ using System.Linq.Expressions;
 using JetBrains.Annotations;
 using OnlineMinion.Contracts.PaymentSpec.Responses;
 using OnlineMinion.Data;
-using OnlineMinion.Data.Entities.Shared;
+using OnlineMinion.Domain.Shared;
 using OnlineMinion.RestApi.Shared.Handlers;
 
 namespace OnlineMinion.RestApi.PaymentSpec.Handlers;
