+++ OnlineMinion.RestApi/PaymentSpec/Handlers/UpdatePaymentSpecReqHlr.cs
@@ -2,7 +2,7 @@ using JetBrains.Annotations;
 using Microsoft.Extensions.Logging;
 using OnlineMinion.Contracts.PaymentSpec.Requests;
 using OnlineMinion.Data;
-using OnlineMinion.Data.Entities.Shared;
+using OnlineMinion.Domain.Shared;
 using OnlineMinion.RestApi.Shared.Handlers;
 
 namespace OnlineMinion.RestApi.PaymentSpec.Handlers;
