+++ OnlineMinion.RestApi/Shared/Handlers/BaseCreateModelReqHlr.cs
@@ -3,6 +3,7 @@ using OnlineMinion.Common;
 using OnlineMinion.Contracts.Shared.Requests;
 using OnlineMinion.Contracts.Shared.Responses;
 using OnlineMinion.Data;
+using OnlineMinion.Domain;
 
 namespace OnlineMinion.RestApi.Shared.Handlers;
