+++ OnlineMinion.RestApi/Shared/Handlers/BaseGetModelByIdReqHlr.cs
@@ -3,6 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using OnlineMinion.Common;
 using OnlineMinion.Contracts.Shared.Requests;
 using OnlineMinion.Data;
+using OnlineMinion.Domain;
 
 namespace OnlineMinion.RestApi.Shared.Handlers;
