+++ OnlineMinion.RestApi/Shared/Handlers/BaseGetSomeModelsPagedReqHlr.cs
@@ -5,6 +5,7 @@ using OnlineMinion.Common;
 using OnlineMinion.Contracts.Shared.Requests;
 using OnlineMinion.Contracts.Shared.Responses;
 using OnlineMinion.Data;
+using OnlineMinion.Domain;
 using OnlineMinion.RestApi.Helpers;
 using OnlineMinion.RestApi.Paging;
