+++ OnlineMinion.RestApi/Shared/Handlers/BaseUpdateModelReqHlr.cs
@@ -3,6 +3,7 @@ using MediatR;
 using Microsoft.Extensions.Logging;
 using OnlineMinion.Contracts.Shared.Requests;
 using OnlineMinion.Data;
+using OnlineMinion.Domain;
 
 namespace OnlineMinion.RestApi.Shared.Handlers;
