+++ OnlineMinion.RestApi/Shared/Handlers/GetSomeModelDescriptorsReqHlr.cs
@@ -5,6 +5,7 @@ using OnlineMinion.Common;
 using OnlineMinion.Contracts;
 using OnlineMinion.Contracts.Shared.Requests;
 using OnlineMinion.Data;
+using OnlineMinion.Domain;
 
 namespace OnlineMinion.RestApi.Shared.Handlers;
