+++ OnlineMinion.RestApi/Transactions/Credit/Handlers/UpdateTransactionCreditReqHlr.cs
@@ -2,7 +2,7 @@ using JetBrains.Annotations;
 using Microsoft.Extensions.Logging;
 using OnlineMinion.Contracts.Transactions.Credit.Requests;
 using OnlineMinion.Data;
-using OnlineMinion.Data.Entities;
+using OnlineMinion.Domain;
 using OnlineMinion.RestApi.Shared.Handlers;
 
 namespace OnlineMinion.RestApi.Transactions.Credit.Handlers;
